Fixed query

+++ src/main/kotlin/nl/blackstardlb/bois/services/models/OutfitMemberStats.kt
@@ -6,7 +6,7 @@ data class OutfitMemberStats(
         val br: Int,
         val prestigeLevel: Int,
         val outfitRank: Int,
-        val minutesPlayed: Long,
+        val minutesPlayed: Double,
         val medicPercentage: Int,
         val killsPM: Double,
         val capturesPH: Double,
